Update openblas to version 0.3.33 / rev 77 via SR 1363694
https://build.opensuse.org/request/show/1363694
by user mslacken + anag_factory
Correct file listing and update-alternative paths

+++ packages/o/openblas/openblas.spec
@@ -24,14 +24,11 @@
 
 %bcond_with ringdisabled
 
-%if 0%{?suse_version} > 1500
-%define a_x _%{_arch}
-%endif
-
 %if "%flavor" == ""
 %define package_name %{pname}
 ExclusiveArch:  do_not_build
 %endif
+
 
 %global build_flags USE_THREAD=1 USE_OPENMP=1
 
@@ -40,6 +37,8 @@ ExclusiveArch:  do_not_build
 %define openblas_so_prio 20
 # we build devel packages only from one flavor
 %define build_devel 1
+ %define flexiname libopenblas.so.%{so_v}
+ %define flexidevelname libopenblas.so
 %endif
 
 %if "%flavor" == "pthreads"
@@ -50,6 +49,8 @@ ExclusiveArch:  do_not_build
  %else
  %define openblas_so_prio 20
  %endif
+ %define flexiname libopenblasp.so.%{so_v}
+ %define flexidevelname libopenblasp.so
 %endif
 
 %if "%flavor" == "openmp"
@@ -59,6 +60,35 @@ ExclusiveArch:  do_not_build
  %define arch_flavor 1
  %define openblas_so_prio 50
  %endif
+ %define flexiname libopenblaso.so.%{so_v}
+ %define flexidevelname libopenblaso.so
+%endif
+# Distribution detection and update-alternatives strategy distroname is for debugging purpose
+%if 0%{?suse_version} <= 1500
+%bcond_without alternatives
+%define distroname LEAP15-SLE15
+# Do not use LIBNAMESUFFIX for new builds as it will not allow
+# the different flavors to be plugin replacements of each other
+%define libnamesuffix LIBNAMESUFFIX=_%{flavor}
+%else
+%define a_x _%{_arch}
+%if 0%{?suse_version} >= 1600 && 0%{?suse_version}  < 1610
+%define distroname LEAP16-SLE16
+%bcond_without alternatives
+%else
+%if 0%{?is_opensuse}
+%if 0%{?suse_version} >= 1699
+%define distroname TW
+%bcond_without alternatives
+%else
+%define distroname LEAP16.1
+%bcond_without alternatives
+%endif
+%else
+%define distroname SLE16.1
+%bcond_with alternatives
+%endif
+%endif
 %endif
 
 %ifarch ppc64le
@@ -92,9 +122,15 @@ ExclusiveArch:  do_not_build
 %define package_name  %{pname}_%{flavor}
 %endif
 %define p_prefix %_prefix
+%if %{with alternatives}
 %define p_includedir %_includedir/%pname
 %define p_libdir %_libdir/openblas%{?flavor:-%{flavor}}
 %define p_testdir %_libexecdir/openblas%{?flavor:-%{flavor}}/tests
+%else
+%define p_includedir %_includedir
+%define p_libdir %_libdir
+%define p_testdir %_libexecdir/openblas/tests
+%endif
 %define p_cmakedir %{p_libdir}/cmake/%{pname}
 %define num_threads 64
 
@@ -105,7 +141,7 @@ ExclusiveArch:  do_not_build
 %endif
 
 Name:           %{package_name}
-Version:        0.3.30
+Version:        0.3.33
 Release:        0
 Summary:        An optimized BLAS library based on GotoBLAS2
 License:        BSD-3-Clause
@@ -125,34 +161,69 @@ BuildRequires:  memory-constraints
 BuildRequires:  gcc%{?cc_v}-fortran
 %endif
 BuildRequires:  gcc-fortran
+%if %{with alternatives}
 BuildRequires:  update-alternatives
 Requires(post): update-alternatives
+Requires(postun): update-alternatives
 Requires(preun): update-alternatives
+%endif
 
 %description
 OpenBLAS is an optimized BLAS library based on GotoBLAS2 1.13 BSD version.
 
-%package     -n lib%{name}%{so_a}
+%package     -n lib%{name}%{so_v}
 Summary:        An optimized BLAS library based on GotoBLAS2, %{flavor} version
 Group:          System/Libraries
-Requires(post): update-alternatives
 Requires(post): coreutils
-Requires(preun): update-alternatives
+%if %{with alternatives}
+Recommends:     compatlib%{name}%{so_v} = %{version}
+%endif
  %if "%flavor" == "serial"
 Obsoletes:      lib%{pname}%{so_v} < %{version}
 Provides:       lib%{pname}%{so_v} = %{version}
  %else
-Obsoletes:      lib%{pname}0
+Obsoletes:      lib%{pname}0 < %{version}
  %endif
  %if "%flavor" == "pthreads"
-Obsoletes:      lib%{pname}p0
+Obsoletes:      lib%{pname}p0 < %{version}
  %endif
  %if "%flavor" == "openmp"
-Obsoletes:      lib%{pname}o0
+Obsoletes:      lib%{pname}o0 < %{version}
  %endif
 
-%description -n lib%{name}%{so_a}
+%description -n lib%{name}%{so_v}
 OpenBLAS is an optimized BLAS library based on GotoBLAS2 1.13 BSD version.
+
+%if %{with alternatives}
+%package     -n compatlib%{name}%{so_a}
+Summary:        Compatiblity library containing OpenBLAS %{flavor} via update-alternatives
+Group:          System/Libraries
+Requires(post): coreutils
+Requires(post): update-alternatives
+Requires(preun): update-alternatives
+Requires:       lib%{name}%{so_a} = %{version}
+Enhances:       update-alternatives
+Obsoletes:      lib%{name}%{so_a} <= 0.3.30
+# Automatically install compatlib when upgrading from >= 0.3.30
+Supplements:    (lib%{name}%{so_a} and update-alternatives)
+
+ %if "%flavor" == "serial"
+Obsoletes:      lib%{pname}%{so_v} < %{version}
+ %else
+Obsoletes:      lib%{pname}0 < %{version}
+ %endif
+ %if "%flavor" == "pthreads"
+Obsoletes:      lib%{pname}p0 < %{version}
+ %endif
+ %if "%flavor" == "openmp"
+Obsoletes:      lib%{pname}o0 < %{version}
+ %endif
+
+%description -n compatlib%{name}%{so_a}
+OpenBLAS for %{flavor} which can replace other BLAS versions with the help of
+update-alternatives.
+Incompatible with transactional enabled OS installations.
+%endif
 
 %package     -n lib%{name}-devel
 Summary:        Development libraries for OpenBLAS, %{flavor} version
@@ -171,6 +242,21 @@ OpenBLAS is an optimized BLAS library based on GotoBLAS2 1.13 BSD version.
 
 This package contains the development libraries for serial OpenBLAS version.
 
+%package     -n compatlib%{name}-devel
+Summary:        Development libraries for OpenBLAS for explicit link to OpenBLAS %{flavor}
+Group:          Development/Libraries/C and C++
+Requires:       %{pname}-common-devel = %{version}
+Requires:       compatlib%{name}%{so_a} = %{version}
+%if 0%{?arch_flavor}
+Provides:       %{pname}-devel = %version
+Provides:       %{pname}-devel(default) = %version
+%else
+Provides:       %{pname}-devel(other) = %version
+%endif
+
+%description -n compatlib%{name}-devel
+Provides the link from lib%{pname}.so to libopenblas.o managed by update-alternatives
+
 %package        devel-static
 Summary:        Static version of OpenBLAS
 Group:          Development/Libraries/C and C++
@@ -184,7 +270,7 @@ This package contains the static libraries.
 %package      -n %{pname}-common-devel
 Summary:        Development headers and libraries for OpenBLAS
 Group:          Development/Libraries/C and C++
-Requires:       (%{pname}-devel(default) or %{pname}-devel(other))
+Requires:       lib%{pname}_serial-devel = %{version}
 Requires(pre):  coreutils
 Requires(post): coreutils
 Obsoletes:      %{pname}-devel < %version
@@ -207,8 +293,7 @@ OpenBLAS is an optimized BLAS library based on GotoBLAS2 1.13 BSD version.
 This package contains test binaries.
 
 %prep
-%setup -q -n OpenBLAS-%{?sha1:%{sha1}}%{!?sha1:%{version}}
-%autopatch -p1
+%autosetup -p1 -n OpenBLAS-%{?sha1:%{sha1}}%{!?sha1:%{version}}
 %ifarch s390
 sed -i -e "s@m32@m31@" Makefile.system
 %endif
@@ -228,6 +313,7 @@ sed -i -e '/^OBJS_EXT+=/s@[^= ]*/test_dgemmt.o *@@' utest/Makefile
 cp %{SOURCE1} .
 
 # create baselibs.conf based on flavor
+%if %{with alternatives}
 cat >  %{_sourcedir}/baselibs.conf <<EOF
 lib%{name}%{so_a}
     +%{p_libdir}/libopenblas.*\.so\.*
@@ -236,8 +322,14 @@ lib%{name}-devel
     requires -%{name}-<targettype>
     requires "lib%{name}%{?so_a}-<targettype> = <version>"
 EOF
+%endif
 
 %build
+# disable lto for i586 and ppc64le, as it causes compilation/linking issues and segfaults
+%ifarch i586 ppc64le
+%global _lto_cflags %{nil}
+%endif
+
 # For static libraries use -ffat-lto-objects to make sure the 'regular'
 # assembler code is generated as well as the intermediate code will be
 # stripped during pre-packaging post-processing. Also, set ldflags_tests
@@ -247,10 +339,6 @@ EOF
 %global ldflags_tests -fno-lto
 %endif
 
-# disable lto for ppc64le, boo#1181733
-%ifarch ppc64le
-%define  _lto_cflags %{nil}
-%endif
 
 # Use DYNAMIC_ARCH everywhere - not sure about PPC?
 # Use DYNAMIC_ARCH to build for multiple targets, use TARGET to specify
@@ -261,6 +349,9 @@ EOF
 %ifarch %ix86 x86_64
 %global openblas_target %openblas_target TARGET=CORE2
 %define openblas_opt BUILD_BFLOAT16=1
+%endif
+%ifarch i586
+%define openblas_opt BINARY=32
 %endif
 %ifarch aarch64
 %global openblas_target %openblas_target TARGET=ARMV8
@@ -299,12 +390,12 @@ EOF
 [[ -z $jobs ]] && jobs=1
 # NEVER use %%_smp_mflags with top level make:
 # set MAKE_NB_JOBS instead and let the build do the work!
-# Do not use LIBNAMESUFFIX for new builds as it will not allow
-# the different flavors to be plugin replacements of each other
-%if 0%{?suse_version} <= 1500
-%define libnamesuffix LIBNAMESUFFIX=_%{flavor}
-%endif
+# On i586, build only 'shared' target to skip tests that segfault
+%ifarch i586
+make shared MAKE_NB_JOBS=$jobs %{?openblas_target} %{?build_flags} \
+%else
 make MAKE_NB_JOBS=$jobs %{?openblas_target} %{?build_flags} \
+%endif
      %{?openblas_opt} \
      COMMON_OPT="%{optflags} %{?addopt}" \
      NUM_THREADS=%{num_threads} V=1 \
@@ -323,7 +414,12 @@ make MAKE_NB_JOBS=$jobs %{?openblas_target} %{?build_flags} \
 # Pass NUM_THREADS again, as it is not propagated from the build step
 # https://github.com/OpenMathLib/OpenBLAS/issues/4275
 mkdir -p %{buildroot}/%{p_testdir}
+# On i586, skip install_tests since tests were not built
+%ifarch i586
+%make_install %{?openblas_target} %{?build_flags} \
+%else
 %make_install install_tests %{?openblas_target} %{?build_flags} \
+%endif
     %{?openblas_opt} \
     NUM_THREADS=%{num_threads} \
     OPENBLAS_LIBRARY_DIR=%{p_libdir} \
@@ -354,11 +450,18 @@ done
 %if 0%{!?build_devel:1}
 # We need the includes only once
 rm -rf %{buildroot}%{p_includedir}/
+# Also remove cmake and pkgconfig files for non-devel flavors when alternatives is OFF
+%if %{without alternatives}
+rm -rf %{buildroot}%{p_cmakedir}/
+rm -rf %{buildroot}%{p_libdir}/pkgconfig/
+%endif
 %endif
 
 # Fix cmake config file
+%if 0%{?build_devel:1}
 sed -i 's|%{buildroot}||g' %{buildroot}%{p_cmakedir}/*.cmake
 sed -i 's|_%{flavor}||g' %{buildroot}%{p_cmakedir}/*.cmake
+%endif
 
 # Remove library type specific so. This is solved differently.
 # Needed when not using LIBNAMESUFFIX.
@@ -389,22 +492,52 @@ find -name \*.f -exec chmod 644 {} +
 #  /usr/lib64/libopenblas.so.<so_v> -> /usr/lib64/openblas_default/libopenblas.so.<so_v>
 #  /usr/lib64/libopenblas.so -> libopenblas.so.<so_v>
 
+%if %{with alternatives}
+# Only create flexiname symlink if it differs from the common library name
+%if "%{?flexiname}" != "" && "%{flexiname}" != "lib%{pname}.so.%{so_v}"
+ln -s %{p_libdir}/lib%{pname}.so.%{so_v} %{buildroot}%{_libdir}/%{flexiname}
+%endif
 install -d %{buildroot}/%{_sysconfdir}/alternatives
 for link in openblas-default libblas.so.3 liblapack.so.3 libcblas.so.3 liblapacke.so.3; do
     ln -s %{_sysconfdir}/alternatives/${link}%{?a_x} %{buildroot}/%{_libdir}/${link}
 done
+%else
+# For older releases without alternatives, create flexiname symlinks only if they differ from the actual filename
+%if "%{?flexiname}" != "" && "%{flexiname}" != "lib%{pname}.so.%{so_v}"
+ln -s lib%{pname}.so.%{so_v} %{buildroot}%{_libdir}/%{flexiname}
+%endif
+%endif
+# compatiblity libraries which are use by flexiblas
 
 %if 0%{?build_devel}
+%if %{with alternatives}
 install -d %{buildroot}%{_libdir}/pkgconfig/
-ln -s %{_sysconfdir}/alternatives/openblas-default%{?a_x}/pkgconfig/openblas.pc %{buildroot}%{_libdir}/pkgconfig/
+ln -s %{p_libdir}/pkgconfig/openblas.pc %{buildroot}%{_libdir}/pkgconfig/openblas.pc
 install -d %{buildroot}/%{_libdir}/cmake
-ln -s %{_sysconfdir}/alternatives/openblas-default%{?a_x}/cmake/openblas %{buildroot}/%{_libdir}/cmake/
+ln -s %{p_libdir}/cmake/openblas %{buildroot}/%{_libdir}/cmake/openblas
+%endif
+%endif
+%if %{with alternatives}
+%{?flexidevelname:ln -s %{p_libdir}/lib%{pname}.so.%{so_v} %{buildroot}%{_libdir}/%{flexidevelname}}
+%else
+# For older releases without alternatives, create flexidevelname symlinks only if they differ
+%if "%{?flexidevelname}" != "" && "%{flexidevelname}" != "lib%{pname}.so"
+ln -s lib%{pname}.so %{buildroot}%{_libdir}/%{flexidevelname}
+%endif
 %endif
 
 # Compatibility Links
 %if 0%{?libnamesuffix:1}
+ %if %{with alternatives}
+# remove installed lib to avoid conflicts even if wit update-alternatives these are ghosted
+rm -f %{buildroot}%{_libdir}/lib%{name}.so.%{so_v}
 ln -s openblas-%{flavor}/lib%{pname}.so.%{so_v} %{buildroot}%{_libdir}/lib%{name}.so.%{so_v}
-ln -s openblas-%{flavor}/lib%{pname}.so %{buildroot}%{_libdir}/lib%{name}.so
+test -e %{buildroot}%{_libdir}/lib%{name}.so || ln -s openblas-%{flavor}/lib%{pname}.so %{buildroot}%{_libdir}/lib%{name}.so
+ %else
+# For older releases without alternatives, symlink directly in same directory
+test -e %{buildroot}%{_libdir}/lib%{name}.so.%{so_v} || ln -s lib%{pname}.so.%{so_v} %{buildroot}%{_libdir}/lib%{name}.so.%{so_v}
+test -e %{buildroot}%{_libdir}/lib%{name}.so || ln -s lib%{pname}.so %{buildroot}%{_libdir}/lib%{name}.so
+ %endif
 %endif
 
 # Ensure directory used in older versions are replaced by symlink properly
@@ -413,7 +546,8 @@ d=%{_libdir}/cmake/openblas
 [ -d $d -a ! -L $d -a "$(rpm -q --qf '%%{NAME}' -f $d 2>/dev/null)" = "openblas-devel" ] \
     && { n=$(mktemp -d $(dirname $d)/tmpd-XXXXX); mv $d $n; rm -rf $n; } || true
 
-%post -n lib%{name}%{so_v}
+%if %{with alternatives}
+%post -n compatlib%{name}%{so_v}
 # There's no way to determine if a setting exists, so just remove it and ignore errors
 %{?a_x:%{_sbindir}/update-alternatives --remove-all openblas-default 2>/dev/null || true}
 %{_sbindir}/update-alternatives --install \
@@ -426,34 +560,66 @@ for lib in libblas.so.3 libcblas.so.3 liblapack.so.3 liblapacke.so.3; do
      %{_libdir}/${lib} ${lib}%{?a_x} %{p_libdir}/lib%{pname}.so.%{so_v}  20
 done
 /sbin/ldconfig
+%endif
 
-%post -n %{pname}-common-devel
-ln -sf lib%{pname}.so.%{so_v} %{_libdir}/lib%{pname}.so
+
+%post -n lib%{name}%{so_v}
+/sbin/ldconfig
 
 %postun -n lib%{name}%{so_v}
+/sbin/ldconfig
+
+%post -n %{pname}-common-devel
+# Always point to serial library by default - this is managed in %%files now, not created in %%post
+
+%if %{with alternatives}
+%postun -n compatlib%{name}%{so_v}
 if [ ! -f %{p_libdir}/lib%{pname}.so.%{so_v} ]; then
+    if [ -e %{_sbindir}/update-alternatives ] ; then
     for lib in libblas.so.3 libcblas.so.3 liblapack.so.3 liblapacke.so.3; do
-	%{_sbindir}/update-alternatives --remove ${lib}%{?a_x} %{_libdir}/lib%{pname}.so.%{so_v}
+	    %{_sbindir}/update-alternatives --remove ${lib}%{?a_x} %{p_libdir}/lib%{pname}.so.%{so_v}
     done
+    fi
 fi
 if [ ! -d %{p_libdir} ]; then
     %{_sbindir}/update-alternatives --remove openblas-default%{?a_x} %{p_libdir}
 fi
+# Remove libopenblas.so.%{so_v} symlink if it's broken
+if [ -h "%_libdir/lib%{pname}.so.%{so_v}" ] && [ ! -e "%_libdir/lib%{pname}.so.%{so_v}" ] ; then
+  rm %_libdir/lib%{pname}.so.%{so_v}
+fi
+# Remove libopenblas.so if it's broken
+if [ -h "%_libdir/libopenblas.so" ] && [ ! -e "%_libdir/libopenblas.so" ] ; then
+  rm %_libdir/libopenblas.so
+fi
 /sbin/ldconfig
+%endif
 
 %check
+%ifnarch i586
 sed -e 's#@FLAVOR@#%{flavor}#' \
     -e 's#@COMPILER@#%{?compiler_family:%compiler_family}}#' \
     -e 's#@DIR@#%{buildroot}/%{p_testdir}#' \
     < %{S:3} > ./openblas_tests_check.sh
 bash ./openblas_tests_check.sh
+%endif
 
 %files -n lib%{name}%{so_a}
 %defattr(-,root,root,-)
-%{p_libdir}/lib%{pname}.so.%{so_v}
+# can be %%{_libdir} so be carefull
+%if %{with alternatives}
 %dir %{p_libdir}
+%endif
+%{p_libdir}/lib%{pname}.so.%{so_v}
 %{?libnamesuffix:%{_libdir}/lib%{name}.so.%{so_v}}
-# Created by %%post
+%if "%{?flexiname}" != "" && "%{flexiname}" != "lib%{pname}.so.%{so_v}"
+%{_libdir}/%{flexiname}
+%endif
+
+%if %{with alternatives}
+%files -n compatlib%{name}%{so_a}
+%defattr(-,root,root,-)
+# Created by %%post with update-alternatives
 %ghost %{_libdir}/lib%{pname}.so.%{so_v}
 %ghost %{_libdir}/openblas-default
 %ghost %{_libdir}/libblas.so.3
@@ -465,14 +631,24 @@ bash ./openblas_tests_check.sh
 %ghost %{_sysconfdir}/alternatives/libcblas.so.3%{?a_x}
 %ghost %{_sysconfdir}/alternatives/liblapack.so.3%{?a_x}
 %ghost %{_sysconfdir}/alternatives/liblapacke.so.3%{?a_x}
+%endif
 
 %files -n lib%{name}-devel
 %{p_libdir}/lib%{pname}.so
 %{?libnamesuffix:%{_libdir}/lib%{name}.so}
-%{p_cmakedir}/
+%if %{with alternatives}
 %dir %{p_libdir}/cmake
+%{p_cmakedir}/
 %dir %{p_libdir}/pkgconfig
-%{p_libdir}/pkgconfig
+%{p_libdir}/pkgconfig/openblas.pc
+# For serial flavor, package the main libopenblas.so symlink
+%if "%{flavor}" == "serial"
+%{_libdir}/lib%{pname}.so
+%endif
+%endif
+%if "%{?flexidevelname}" != "" && "%{flexidevelname}" != "lib%{pname}.so"
+%{_libdir}/%{flexidevelname}
+%endif
 
 %files tests
 %dir %{p_testdir}
@@ -486,8 +662,6 @@ bash ./openblas_tests_check.sh
 %files -n %{pname}-common-devel
 %license LICENSE
 %doc Changelog.txt GotoBLAS* README.md README.SUSE
-# created by %%post
-%ghost %{_libdir}/lib%{pname}.so
 %{p_includedir}/
 %{_libdir}/pkgconfig/openblas.pc
 %dir %{_libdir}/cmake
